pull requeue function into projects
