+++ express-api/src/services/projects/projectsServices.ts
@@ -22,6 +22,7 @@ import {
   In,
   InsertResult,
   IsNull,
+  MoreThan,
   Not,
   QueryRunner,
 } from 'typeorm';
@@ -30,11 +31,13 @@ import { PropertyType } from '@/constants/propertyType';
 import { ProjectRisk } from '@/constants/projectRisk';
 import notificationServices, {
   AgencyResponseType,
+  NotificationAudience,
   NotificationStatus,
 } from '../notifications/notificationServices';
 import {
   constructFindOptionFromQuery,
   constructFindOptionFromQuerySingleSelect,
+  getDaysBetween,
 } from '@/utilities/helperFunctions';
 import { ProjectTimestamp } from '@/typeorm/Entities/ProjectTimestamp';
 import { ProjectMonetary } from '@/typeorm/Entities/ProjectMonetary';
@@ -44,6 +47,7 @@ import { ProjectJoin } from '@/typeorm/Entities/views/ProjectJoinView';
 import { Roles } from '@/constants/roles';
 import { PimsRequestUser } from '@/middleware/userAuthCheck';
 import { User } from '@/typeorm/Entities/User';
+import { ProjectStatusNotification } from '@/typeorm/Entities/ProjectStatusNotification';
 
 const projectRepo = AppDataSource.getRepository(Project);
 
@@ -1121,6 +1125,109 @@ const getProjectsForExport = async (filter: ProjectFilter) => {
   }));
 };
 
+const queueOutstandingERPNotifications = async (project: Project, agency: Agency, user: User) => {
+  const queryRunner = AppDataSource.createQueryRunner();
+
+  try {
+    await queryRunner.startTransaction();
+    const projectAgency = await queryRunner.manager.findOne(Agency, {
+      where: { Id: project.AgencyId },
+    });
+    const projectNotes = await queryRunner.manager.find(ProjectNote, {
+      where: { ProjectId: project.Id },
+    });
+    project.Agency = projectAgency;
+    project.Notes = projectNotes;
+
+    // Deal with this project's notifications. Only the 30, 60, 90 days notifications must be queued.
+    // NOTE: Do not use handleProjectNotifcations
+    const notifsToSend: NotificationQueue[] = [];
+    // If this agency owns this project, queue those notifications
+    if (project.AgencyId == agency.Id) {
+      const dateInERP = project.ApprovedOn;
+      const daysSinceThisStatus = getDaysBetween(dateInERP, new Date());
+      const projectStatusNotifs = await queryRunner.manager.find(ProjectStatusNotification, {
+        where: {
+          ToStatusId: ProjectStatus.APPROVED_FOR_ERP,
+          DelayDays: MoreThan(daysSinceThisStatus),
+          Template: {
+            Audience: NotificationAudience.OwningAgency,
+          },
+        },
+        relations: { Template: true },
+      });
+      await Promise.all(
+        projectStatusNotifs.map(async (n) => {
+          //If there is already a pending notification for this agency with this template, skip.
+          const notifExists = await queryRunner.manager.exists(NotificationQueue, {
+            where: [
+              {
+                ProjectId: project.Id,
+                ToAgencyId: agency.Id,
+                TemplateId: n.TemplateId,
+                Status: NotificationStatus.Accepted,
+              },
+              {
+                ProjectId: project.Id,
+                ToAgencyId: agency.Id,
+                TemplateId: n.TemplateId,
+                Status: NotificationStatus.Pending,
+              },
+            ],
+          });
+
+          if (!notifExists) {
+            // If there is no notification like this already pending, we send one.
+            const sendOn = new Date(dateInERP.getTime());
+            sendOn.setDate(dateInERP.getDate() + n.DelayDays);
+            notifsToSend.push(
+              await notificationServices.insertProjectNotificationQueue(
+                n.Template,
+                n,
+                project,
+                agency,
+                undefined,
+                sendOn,
+                queryRunner,
+              ),
+            );
+          }
+        }),
+      );
+    } else {
+      // Otherwise, handle watch notifications for this agency
+      // Update response
+      const updatedResponse: Partial<ProjectAgencyResponse> = {
+        Response: AgencyResponseType.Subscribe,
+        AgencyId: agency.Id,
+        ProjectId: project.Id,
+        OfferAmount: 0,
+        CreatedById: user.Id,
+      };
+      const savedResponse = await queryRunner.manager.save(ProjectAgencyResponse, updatedResponse);
+      const watchNotifications = await notificationServices.generateProjectWatchNotifications(
+        project,
+        [savedResponse],
+        queryRunner,
+      );
+      notifsToSend.push(...watchNotifications);
+    }
+
+    const result = await Promise.all(
+      notifsToSend.map((notif) => notificationServices.sendNotification(notif, user, queryRunner)),
+    );
+    await queryRunner.commitTransaction();
+    return result;
+  } catch {
+    await queryRunner.rollbackTransaction();
+    return new Error(
+      `Failed to queue notifications for project ID ${project.Id} when updating agency ID ${agency.Id}`,
+    );
+  } finally {
+    await queryRunner.release();
+  }
+};
+
 const projectServices = {
   addProject,
   getProjectById,
@@ -1129,6 +1236,7 @@ const projectServices = {
   getProjects,
   getProjectsForExport,
   handleProjectNotifications,
+  queueOutstandingERPNotifications,
 };
 
 export default projectServices;
